RandomMoveLocEffect/TeleportService byte<->sbyte: (byte) cast GetH() heading arg to World.UpdatePosition + TeleportTo (faithful Java byte=sbyte; GetH() sbyte->byte param); clears 2 CS1503 (891->889)

+++ dotnetConversion/src/Aion.GameServer/Services/Teleport/TeleportService.cs
@@ -428,7 +428,7 @@ public class TeleportService
         InstanceExit instanceExit = DataManager.INSTANCE_EXIT_DATA.GetInstanceExit(worldId, race);
         if (instanceExit != null && InstanceService.InstanceExists(instanceExit.GetExitWorld(), 1))
         {
-            TeleportTo(player, instanceExit.GetExitWorld(), instanceExit.GetX(), instanceExit.GetY(), instanceExit.GetZ(), instanceExit.GetH());
+            TeleportTo(player, instanceExit.GetExitWorld(), instanceExit.GetX(), instanceExit.GetY(), instanceExit.GetZ(), (byte)instanceExit.GetH());
         }
         else
         {
